backend/requirements.txt
# MarketNavigator v2 - Backend Requirements (High-Scale Production)

# Django Core
Django>=5.0,<5.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
djangorestframework-simplejwt>=5.3.0
django-filter>=23.5

# Async & WebSocket
daphne>=4.1.0
channels>=4.0.0
channels-redis>=4.2.0

# Database
psycopg2-binary>=2.9.9
django-db-connection-pool>=1.2.4

# Caching (Redis)
django-redis>=5.4.0
redis>=5.0.1
hiredis>=2.3.2

# Task Queue
celery>=5.3.4
flower>=2.0.1

# AI Integration
openai>=1.10.0
tiktoken>=0.5.2

# HTTP Clients (with retry support)
httpx>=0.26.0
aiohttp>=3.9.0
tenacity>=8.2.3

# AWS S3 / Object Storage
boto3>=1.34.0
django-storages>=1.14.2

# File handling
python-magic>=0.4.27
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
python-decouple>=3.8

# Rate Limiting
django-ratelimit>=4.1.0

# API Documentation
drf-spectacular>=0.27.0

# Monitoring & Logging
sentry-sdk>=1.39.0

# Server
gunicorn>=21.0.0
uvicorn>=0.27.0

# Data processing
pandas>=2.1.0

# Markdown to HTML for reports
markdown>=3.5.0
